fix cart :add price

+++ model/cart_model.py
@@ -8,7 +8,7 @@ class CartCreate(BaseModel):
     user_id: int
     product_id: UUID
     quantity: int = 1
-    price:int
+    price:float
 
 
 class CartUpdate(BaseModel):
@@ -20,7 +20,7 @@ class CartResponse(BaseModel):
     user_id: int
     product_id: UUID
     quantity: int
-    price:int
+    price:float
     created_at: datetime
 
     class Config:
